# API/PHP/Laravel/lib/app/Models/CreateUserForm.php
<?php
/**
 * CreateUserForm
 */
namespace app\Models;

/**
 * CreateUserForm
 */
class CreateUserForm {

    /** @var string $email */
    private $email;

    /** @var string $password */
    private $password;

    /** @var string $username */
    private $username;

    /** @var \app\Models\CreateUserFormFullName $fullName */
    private $fullName;

}